pyproject.toml
[project]
name = "people-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Max Harley", email = "t94j0@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "logfire>=3.1.1",
    "pydantic-ai>=0.0.19",
    "tavily-python>=0.5.0",
]

[dependency-groups]
lint = [
    "ruff>=0.9.2",
]

[project.scripts]
people-researcher = "people_researcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
extend-select = ["Q", "RUF100", "C90", "UP", "I", "D"]
isort = { combine-as-imports = true, known-first-party = ["pydantic_ai"] }
mccabe = { max-complexity = 15 }
ignore = [
    "D100", # ignore missing docstring in module
    "D102", # ignore missing docstring in public method
    "D104", # ignore missing docstring in public package
    "D105", # ignore missing docstring in magic methods
    "D107", # ignore missing docstring in __init__ methods
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]

[tool.pyright]
typeCheckingMode = "strict"
reportUnnecessaryTypeIgnoreComment = true
reportMissingTypeStubs = false
venvPath = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
